comment and likes finished

+++ backend/src/main/java/com/blog/service/BlogService.java
@@ -10,9 +10,9 @@ import com.blog.dto.BlogUpdateRequest;
 @Service
 public interface BlogService {
     List<BlogResponse> getAllBlogs();
-    BlogResponse getBlogDetails(Integer id);
+    BlogResponse getBlogDetails(Long id);
     BlogResponse createBlog(BlogRequest request, String username);
-    BlogResponse updateBlog(Integer id, BlogUpdateRequest request,String username);
-    void deleteBlog(Integer id, String username);
+    BlogResponse updateBlog(Long id, BlogUpdateRequest request,String username);
+    void deleteBlog(Long id, String username);
     List<BlogResponse> getBlogsByUser(String username);
 }
